Refactored SR5Item.prepareData similar to SR5Actor.prepareData
